Refactor Home Assistant sensor modules
